+++ AmazonChimeSDK/AmazonChimeSDK/internal/video/VideoClientProtocol.swift
@@ -51,8 +51,6 @@ import Foundation
 
     func demoteFromPrimaryMeeting()
 
-    func setSimulcast(_ simulcast: Bool)
-
     func setMaxBitRateKbps(_ maxBitRate: UInt32)
 
     func setContentMaxResolutionUHD(_ isContentMaxResolutionUHD: Bool)